Create blackjack.cpp
Plays the classic game of blackjack, without the gambling, against a computer.

Miscellaneous:
To run the file, a number must be passed in to seed the random shuffling. For example, running ./blackjack 15 in the command line will set the seed to 15.
After the game has concluded, the choice to play again or quit will appear. Typing in 'y' will continue the game, any other input will terminate the program.